dubbo-samples-rest:change groupId and package to org.apache.dubbo
