--- dubbo-samples-rest/src/main/java/com/alibaba/dubbo/samples/rest/api/extension/ClientTraceFilter.java
@@ -1,38 +1,0 @@
-/*
- *
- *   Licensed to the Apache Software Foundation (ASF) under one or more
- *   contributor license agreements.  See the NOTICE file distributed with
- *   this work for additional information regarding copyright ownership.
- *   The ASF licenses this file to You under the Apache License, Version 2.0
- *   (the "License"); you may not use this file except in compliance with
- *   the License.  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- *
- */
-package com.alibaba.dubbo.samples.rest.api.extension;
-
-import java.io.IOException;
-
-import javax.ws.rs.client.ClientRequestContext;
-import javax.ws.rs.client.ClientRequestFilter;
-import javax.ws.rs.client.ClientResponseContext;
-import javax.ws.rs.client.ClientResponseFilter;
-
-public class ClientTraceFilter implements ClientRequestFilter, ClientResponseFilter {
-
-    public void filter(ClientRequestContext requestContext) throws IOException {
-        System.out.println("Client request filter invoked");
-    }
-
-    public void filter(ClientRequestContext clientRequestContext, ClientResponseContext clientResponseContext) throws
-        IOException {
-        System.out.println("Client response filter invoked");
-    }
-}
