--- dubbo-samples-rest/src/main/java/com/alibaba/dubbo/samples/rest/api/extension/CustomExceptionMapper.java
@@ -1,34 +1,0 @@
-/*
- *
- *   Licensed to the Apache Software Foundation (ASF) under one or more
- *   contributor license agreements.  See the NOTICE file distributed with
- *   this work for additional information regarding copyright ownership.
- *   The ASF licenses this file to You under the Apache License, Version 2.0
- *   (the "License"); you may not use this file except in compliance with
- *   the License.  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- *
- */
-package com.alibaba.dubbo.samples.rest.api.extension;
-
-import org.apache.dubbo.rpc.RpcContext;
-
-import javax.ws.rs.NotFoundException;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.ext.ExceptionMapper;
-
-public class CustomExceptionMapper implements ExceptionMapper<NotFoundException> {
-
-    public Response toResponse(NotFoundException e) {
-        System.out.println("Exception mapper successfully got an exception: " + e + ":" + e.getMessage());
-        System.out.println("Client IP is " + RpcContext.getContext().getRemoteAddressString());
-        return Response.status(Response.Status.NOT_FOUND).entity("Oops! the requested resource is not found!").type("text/plain").build();
-    }
-}
