--- dubbo-samples-rest/src/main/java/com/alibaba/dubbo/samples/rest/api/facade/AnotherUserRestService.java
@@ -1,45 +1,0 @@
-/*
- *
- *   Licensed to the Apache Software Foundation (ASF) under one or more
- *   contributor license agreements.  See the NOTICE file distributed with
- *   this work for additional information regarding copyright ownership.
- *   The ASF licenses this file to You under the Apache License, Version 2.0
- *   (the "License"); you may not use this file except in compliance with
- *   the License.  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- *
- */
-package com.alibaba.dubbo.samples.rest.api.facade;
-
-import javax.validation.constraints.Min;
-import javax.ws.rs.Consumes;
-import javax.ws.rs.GET;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
-
-import com.alibaba.dubbo.samples.rest.api.User;
-import com.alibaba.dubbo.rpc.protocol.rest.support.ContentType;
-
-@Path("u")
-@Consumes({MediaType.APPLICATION_JSON, MediaType.TEXT_XML})
-@Produces({ContentType.APPLICATION_JSON_UTF_8, ContentType.TEXT_XML_UTF_8})
-public interface AnotherUserRestService {
-
-    @GET
-    @Path("{id : \\d+}")
-    User getUser(@PathParam("id") @Min(1L) Long id);
-
-    @POST
-    @Path("register")
-    RegistrationResult registerUser(User user);
-}
