--- dubbo-samples-rest/src/main/java/com/alibaba/dubbo/samples/rest/impl/facade/AnotherUserRestServiceImpl.java
@@ -1,51 +1,0 @@
-/*
- *
- *   Licensed to the Apache Software Foundation (ASF) under one or more
- *   contributor license agreements.  See the NOTICE file distributed with
- *   this work for additional information regarding copyright ownership.
- *   The ASF licenses this file to You under the Apache License, Version 2.0
- *   (the "License"); you may not use this file except in compliance with
- *   the License.  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- *
- */
-package com.alibaba.dubbo.samples.rest.impl.facade;
-
-import com.alibaba.dubbo.samples.rest.api.User;
-import com.alibaba.dubbo.samples.rest.api.UserService;
-import com.alibaba.dubbo.samples.rest.api.facade.AnotherUserRestService;
-import com.alibaba.dubbo.samples.rest.api.facade.RegistrationResult;
-
-import org.apache.dubbo.rpc.RpcContext;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service("anotherUserRestService")
-public class AnotherUserRestServiceImpl implements AnotherUserRestService {
-
-    @Autowired
-    private UserService userService;
-
-    public void setUserService(UserService userService) {
-        this.userService = userService;
-    }
-
-    @Override
-    public User getUser(Long id) {
-        System.out.println("Client name is " + RpcContext.getContext().getAttachment("clientName"));
-        System.out.println("Client impl is " + RpcContext.getContext().getAttachment("clientImpl"));
-        return userService.getUser(id);
-    }
-
-    @Override
-    public RegistrationResult registerUser(User user) {
-        return new RegistrationResult(userService.registerUser(user));
-    }
-}
