--- dubbo-samples-rest/src/main/java/com/alibaba/dubbo/samples/rest/impl/facade/UserRestServiceImpl.java
@@ -1,78 +1,0 @@
-/*
- *
- *   Licensed to the Apache Software Foundation (ASF) under one or more
- *   contributor license agreements.  See the NOTICE file distributed with
- *   this work for additional information regarding copyright ownership.
- *   The ASF licenses this file to You under the Apache License, Version 2.0
- *   (the "License"); you may not use this file except in compliance with
- *   the License.  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- *
- */
-package com.alibaba.dubbo.samples.rest.impl.facade;
-
-import com.alibaba.dubbo.rpc.protocol.rest.support.ContentType;
-import com.alibaba.dubbo.samples.rest.api.User;
-import com.alibaba.dubbo.samples.rest.api.UserService;
-import com.alibaba.dubbo.samples.rest.api.facade.RegistrationResult;
-import com.alibaba.dubbo.samples.rest.api.facade.UserRestService;
-
-import org.apache.dubbo.rpc.RpcContext;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.ws.rs.Consumes;
-import javax.ws.rs.GET;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
-
-@Service("userRestService")
-@Path("users")
-@Consumes({MediaType.APPLICATION_JSON, MediaType.TEXT_XML})
-@Produces({ContentType.APPLICATION_JSON_UTF_8, ContentType.TEXT_XML_UTF_8})
-public class UserRestServiceImpl implements UserRestService {
-
-//    private static final Logger logger = LoggerFactory.getLogger(UserRestServiceImpl.class);
-
-    @Autowired
-    private UserService userService;
-
-    public void setUserService(UserService userService) {
-        this.userService = userService;
-    }
-
-    @Override
-    @GET
-    @Path("{id : \\d+}")
-    public User getUser(@PathParam("id") Long id/*, @Context HttpServletRequest request*/) {
-        // test context injection
-//        System.out.println("Client address from @Context injection: " + (request != null ? request.getRemoteAddr() : ""));
-        System.out.println("Client address from RpcContext: " + RpcContext.getContext().getRemoteAddressString());
-        if (RpcContext.getContext().getRequest(HttpServletRequest.class) != null) {
-            System.out.println("Client IP address from RpcContext: " + RpcContext.getContext().getRequest(HttpServletRequest.class).getRemoteAddr());
-        }
-        if (RpcContext.getContext().getResponse(HttpServletResponse.class) != null) {
-            System.out.println("Response object from RpcContext: " + RpcContext.getContext().getResponse(HttpServletResponse.class));
-        }
-        return userService.getUser(id);
-    }
-
-    @Override
-    @POST
-    @Path("register")
-    public RegistrationResult registerUser(User user) {
-        return new RegistrationResult(userService.registerUser(user));
-    }
-}
