+++ dubbo-samples-rest/src/main/java/org/apache/dubbo/samples/rest/api/extension/ClientTraceFilter.java
@@ -1,0 +1,37 @@
+/*
+ *
+ *   Licensed to the Apache Software Foundation (ASF) under one or more
+ *   contributor license agreements.  See the NOTICE file distributed with
+ *   this work for additional information regarding copyright ownership.
+ *   The ASF licenses this file to You under the Apache License, Version 2.0
+ *   (the "License"); you may not use this file except in compliance with
+ *   the License.  You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ *
+ */
+package org.apache.dubbo.samples.rest.api.extension;
+
+import javax.ws.rs.client.ClientRequestContext;
+import javax.ws.rs.client.ClientRequestFilter;
+import javax.ws.rs.client.ClientResponseContext;
+import javax.ws.rs.client.ClientResponseFilter;
+
+public class ClientTraceFilter implements ClientRequestFilter, ClientResponseFilter {
+
+    @Override
+    public void filter(ClientRequestContext requestContext) {
+        System.out.println("Client request filter invoked");
+    }
+
+    @Override
+    public void filter(ClientRequestContext clientRequestContext, ClientResponseContext clientResponseContext) {
+        System.out.println("Client response filter invoked");
+    }
+}
