+++ dubbo-samples-rest/src/main/resources/spring/rest-consumer.xml
@@ -29,9 +29,9 @@
     <dubbo:registry address="zookeeper://127.0.0.1:2181"/>
 
     <dubbo:reference id="anotherUserRestService"
-                     interface="com.alibaba.dubbo.samples.rest.api.facade.AnotherUserRestService"/>
+                     interface="org.apache.dubbo.samples.rest.api.facade.AnotherUserRestService"/>
 
     <!-- directly connect to provider to simulate the access to non-dubbo rest services -->
-    <!--<dubbo:reference id="anotherUserRestService" interface="com.alibaba.dubbo.demo.user.facade.AnotherUserRestService" url="rest://localhost:8888/services/"/>-->
+    <!--<dubbo:reference id="anotherUserRestService" interface="org.apache.dubbo.demo.user.facade.AnotherUserRestService" url="rest://localhost:8888/services/"/>-->
 
 </beans>
