+++ dubbo-samples-rest/src/main/resources/spring/rest-provider.xml
@@ -25,7 +25,7 @@
        xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
 	http://code.alibabatech.com/schema/dubbo http://code.alibabatech.com/schema/dubbo/dubbo.xsd http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context.xsd">
 
-    <context:component-scan base-package="com.alibaba.dubbo.samples.rest"/>
+    <context:component-scan base-package="org.apache.dubbo.samples.rest"/>
 
     <dubbo:application name="rest-provider" owner="programmer" organization="dubbo"/>
 
@@ -36,34 +36,34 @@
 
     <!-- use tomcat server -->
     <dubbo:protocol name="rest" port="8888" threads="500" contextpath="services" server="tomcat" accepts="500"
-                    extension="com.alibaba.dubbo.samples.rest.api.extension.TraceInterceptor,
-                    com.alibaba.dubbo.samples.rest.api.extension.TraceFilter,
-                    com.alibaba.dubbo.samples.rest.api.extension.ClientTraceFilter,
-                    com.alibaba.dubbo.samples.rest.api.extension.DynamicTraceBinding,
-                    com.alibaba.dubbo.samples.rest.api.extension.CustomExceptionMapper,
-                    com.alibaba.dubbo.rpc.protocol.rest.support.LoggingFilter"/>
+                    extension="org.apache.dubbo.samples.rest.api.extension.TraceInterceptor,
+                    org.apache.dubbo.samples.rest.api.extension.TraceFilter,
+                    org.apache.dubbo.samples.rest.api.extension.ClientTraceFilter,
+                    org.apache.dubbo.samples.rest.api.extension.DynamicTraceBinding,
+                    org.apache.dubbo.samples.rest.api.extension.CustomExceptionMapper,
+                    org.apache.dubbo.rpc.protocol.rest.support.LoggingFilter"/>
 
     <dubbo:protocol name="dubbo" port="20880" threads="200"/>
 
     <!-- use the external tomcat or other server with the servlet approach; the port and contextpath must be exactly the same as those in external server -->
     <!--<dubbo:protocol name="rest" port="8888" contextpath="services" server="servlet"/>-->
 
-    <!-- we add the group property since there's another annotation-configured service impl: com.alibaba.dubbo.demo.user.facade.AnnotationDrivenUserRestServiceImpl -->
-    <dubbo:service interface="com.alibaba.dubbo.samples.rest.api.UserService" ref="userService" protocol="dubbo"
+    <!-- we add the group property since there's another annotation-configured service impl: org.apache.dubbo.demo.user.facade.AnnotationDrivenUserRestServiceImpl -->
+    <dubbo:service interface="org.apache.dubbo.samples.rest.api.UserService" ref="userService" protocol="dubbo"
                    group="xmlConfig"/>
 
-    <dubbo:service interface="com.alibaba.dubbo.samples.rest.api.facade.UserRestService" ref="userRestService"
+    <dubbo:service interface="org.apache.dubbo.samples.rest.api.facade.UserRestService" ref="userRestService"
                    protocol="rest" validation="true"/>
-    <dubbo:service interface="com.alibaba.dubbo.samples.rest.api.facade.AnotherUserRestService"
+    <dubbo:service interface="org.apache.dubbo.samples.rest.api.facade.AnotherUserRestService"
                    ref="anotherUserRestService" protocol="rest,dubbo" timeout="2000" connections="100"
                    validation="true"/>
 
 
     <!-- SwaggerUI -->
-    <bean id="swaggerService" class="com.alibaba.dubbo.integration.swagger.DubboSwaggerApiListingResource" />
+    <bean id="swaggerService" class="org.apache.dubbo.integration.swagger.DubboSwaggerApiListingResource" />
     <bean id="beanConfig" class="io.swagger.jaxrs.config.BeanConfig">
         <property name="schemes" value="http" />
-        <property name="resourcePackage" value="com.alibaba.dubbo.samples.rest.api"/>
+        <property name="resourcePackage" value="org.apache.dubbo.samples.rest.api"/>
         <property name="version" value="2.0"/>
         <property name="host" value="localhost:8888"/>
         <property name="basePath" value="/services/"/>
@@ -74,7 +74,7 @@
         <property name="licenseUrl" value="http://www.apache.org/licenses/LICENSE-2.0.html"/>
         <property name="scan" value="true" />
     </bean>
-    <dubbo:service interface="com.alibaba.dubbo.integration.swagger.DubboSwaggerService" ref="swaggerService" protocol="rest" />
+    <dubbo:service interface="org.apache.dubbo.integration.swagger.DubboSwaggerService" ref="swaggerService" protocol="rest" />
 
 
 </beans>
